Inline canvas utils that are only used once
