+++ src/layout/Surface.js
@@ -3,10 +3,37 @@
 import type {Interaction} from '../useCanvasInteraction';
 import type {Size} from './geometry';
 
-import {getCanvasContext} from '../content-views/canvasUtils';
+import memoize from 'memoize-one';
 
 import {View} from './View';
 import {zeroPoint} from './geometry';
+
+// hidpi canvas: https://www.html5rocks.com/en/tutorials/canvas/hidpi/
+function configureRetinaCanvas(canvas, height, width) {
+  const dpr: number = window.devicePixelRatio || 1;
+  canvas.width = width * dpr;
+  canvas.height = height * dpr;
+  canvas.style.width = `${width}px`;
+  canvas.style.height = `${height}px`;
+  return dpr;
+}
+
+const getCanvasContext = memoize(
+  (
+    canvas: HTMLCanvasElement,
+    height: number,
+    width: number,
+    scaleCanvas: boolean = true,
+  ): CanvasRenderingContext2D => {
+    const context = canvas.getContext('2d', {alpha: false});
+    if (scaleCanvas) {
+      const dpr = configureRetinaCanvas(canvas, height, width);
+      // Scale all drawing operations by the dpr, so you don't have to worry about the difference.
+      context.scale(dpr, dpr);
+    }
+    return context;
+  },
+);
 
 /**
  * Represents the canvas surface and a view heirarchy. A surface is also the
